bootstrap: error when no snapshot is received
when theere are connection issues, no snapshot is received. in
those cases we want to refuse the start of the algo order with an
error, because we don't know what state our account is in.

this error can be handled by the spawning server, but should be
an exception. the server should run `connect` on the host and
then wait for the ready event to get emitted, before starting the
first algo order.