ws-api: handle unknown errors
this catches all unknown errors returned by the api, so the algo
order is shut down properly and does not remain in idle state.

in case there is an api error returned with "action disabled"
the algo order would keep running, no atomic orders would get
submitted and no error would be shown.

this PR handles unknown error responses and sends an
error message to the ui. it also stops the algo order that errors