ma-cross: guard against outdated data (#125)
* don't execute on historic candles

the ws api sends a snapshot in the beginning. it contains historic
candles from the past. additionally, when there is a lot of traffic,
the ws2 sends the last 1-3 candles that are not part of the
snapshot yet.

example, on a 5min scale:

last candle in snapshot: 13:30
immediately sent after connect: candles for 13:35 and 13:40
current date time: 13:41

this issue leads to an issue where, when with those 1-3 historic
candles a ma cross happens, it is executed - but too late.

to fix this, we get the current server time on start of the algo,
and then check if the incoming data is older.

related PRs:

https://github.com/bitfinexcom/bfx-hf-algo/pull/128
https://github.com/bitfinexcom/bfx-api-node-core/pull/18