storage: apply changes to receive algo order data
to support worker threads and in general better support for
hosted environments with many clients, storage won't be done
every few seconds any more, like currently the case

on start the state is stored, and on shutdown. this reduces the
db writes a lot (a twap with a 2 secod interval currently writes
every 2 seconds). it also brings flexibility by loosening the
coupling to the hf-models and database adapters, which we won't
use for the hosted version.

the event `ao:persist:db:update` is kept as `ao:state:update` and
can get used for detailed UI updates on algo order details in the
future